requirements.txt
dj-database-url==0.4.1
Django==1.10
requests==2.12.4
psycopg2==2.6.2
python-social-auth==0.2.21
gunicorn==19.6.0
whitenoise==3.2.1
pygithub==1.26.0
django-rq==0.9.2
raven==5.32.0
